+++ lib/widgets/weather/weather_floating.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 
 import 'package:provider/provider.dart';
 
-import 'package:smartclock/widgets/weather/icons.dart';
+import 'package:smartclock/widgets/weather/weather_icons.dart';
 import 'package:smartclock/config/config.dart' show ConfigModel, Config;
 
 class WeatherFloating extends StatelessWidget {
